Move functions from __init__ to utils, fix a few bugs with db references
