+++ backend/cloud_inquisitor/plugins/types/issues.py
@@ -172,11 +172,11 @@ class BaseIssue(ABC):
         Returns:
             list of issue objects
         """
-        qry = db.Issue.find(
+        issues = db.Issue.find(
             Issue.issue_type_id == IssueType.get(cls.issue_type).issue_type_id
         )
 
-        return {res.issue_id: cls(res) for res in qry.all()}
+        return {res.issue_id: cls(res) for res in issues}
 
     @classmethod
     def search(cls, *, limit=100, page=1, properties=None, return_query=False):
